add support for customer and tenant users
TB API is a little tricky as it provides different API variants for each
type of user even when they do exactly the same thing. Detect which type
of user we are and use the proper API to list assets and devices.

Public is a beast of its own, it's a customer but without credentials so
it needs a dedicated login entry point. Also if you try to get
informations about him with /api/auth/user it won't get you anything
while it works with proper customers.