save refresh token
if download takes a while you may get a 401 unauthorized errors because
the login token is expired. Calling refresh once in a while can prevent
this. At some point I'd like to turn this into a context manager that
takes care of periodically refreshing the session. Let's do it manually
for now.